+++ Tests/PackageCollectionsSigningTests/SigningTests+ECKey.swift
@@ -15,6 +15,8 @@ import XCTest
 
 class ECKeySigningTests: XCTestCase {
     func test_signAndValidate_happyCase() throws {
+        try skipIfUnsupportedPlatform()
+
         let privateKey = try ECPrivateKey(pem: ecPrivateKey.bytes)
         let publicKey = try ECPublicKey(pem: ecPublicKey.bytes)
 
@@ -24,6 +26,8 @@ class ECKeySigningTests: XCTestCase {
     }
 
     func test_signAndValidate_mismatch() throws {
+        try skipIfUnsupportedPlatform()
+
         let privateKey = try ECPrivateKey(pem: ecPrivateKey.bytes)
         let publicKey = try ECPublicKey(pem: ecPublicKey.bytes)