+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Swift v.Next
 
     Adding a dependency requirement can now be done with the convenience initializer `.package(url: String, revision: String)`.
 
+* [#3292]
+   * Improvements
+   
+   Adding a dependency requirement can now be done with the convenience initializer `.package(url: String, branch: String)`.
+
+
 
 Swift 5.4
 -----------
@@ -29,6 +35,8 @@ Swift 5.4
   * Source Breakages for Swift Packages
     
     The package manager now throws an error if a manifest file contains invalid UTF-8 byte sequences.
+    
+
 
 Swift 4.2
 ---------